separated pages/components for provider side
